store part 6

+++ api/src/main/java/com/onlineMarket/api/dto/OrderDto.java
@@ -1,12 +1,13 @@
 package com.onlineMarket.api.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class OrderDto {
     private Long id;
     private String username;
     private List<CartItemDto> items;
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
     private String address;
     private String phone;
 
@@ -34,11 +35,11 @@ public class OrderDto {
         this.items = items;
     }
 
-    public Integer getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
@@ -58,7 +59,7 @@ public class OrderDto {
         this.phone = phone;
     }
 
-    public OrderDto(Long id, String username, List<CartItemDto> items, Integer totalPrice, String address, String phone) {
+    public OrderDto(Long id, String username, List<CartItemDto> items, BigDecimal totalPrice, String address, String phone) {
         this.id = id;
         this.username = username;
         this.items = items;
